🚚 [projects] Move function `skipupdate` from `services` to `update` module
