+++ src/cutty/services/update.py
@@ -44,14 +44,6 @@ def update(
     updateproject(projectdir, createproject)
 
 
-def skipupdate(projectdir: Path) -> None:
-    """Skip an update with conflicts."""
-    project = Repository.open(projectdir)
-    project.resetcherrypick()
-
-    project.heads[LATEST_BRANCH] = project.heads[UPDATE_BRANCH]
-
-
 def abortupdate(projectdir: Path) -> None:
     """Abort an update with conflicts."""
     project = Repository.open(projectdir)
